[cleanup] extract logic from App to AuthContext file
