+++ src/screens/auth/AuthContext.ts
@@ -1,0 +1,106 @@
+import AsyncStorage from '@react-native-async-storage/async-storage';
+import {
+  createUserWithEmailAndPassword,
+  getAuth,
+  signInWithEmailAndPassword,
+} from 'firebase/auth';
+import { createContext, useReducer } from 'react';
+
+export type AuthContextType = {
+  signIn: (email: string, password: string) => void;
+  signOut: () => void;
+  signUp: (email: string, password: string) => void;
+  authState?: AuthState;
+};
+
+export type AuthState = {
+  userToken: string | null;
+  isLoading: boolean;
+  isSignout: boolean; // TODO: @wangannie use this to change the animation of the screen when signing out
+};
+
+export const AuthContext = createContext<AuthContextType>(
+  {} as AuthContextType,
+);
+
+export type AuthContextAction =
+  | { type: 'RESTORE_TOKEN'; token: string | null }
+  | { type: 'SIGN_IN'; token: string }
+  | { type: 'SIGN_OUT' };
+
+export const useAuthReducer = () =>
+  useReducer(
+    (prevState: AuthState, action: AuthContextAction) => {
+      switch (action.type) {
+        case 'RESTORE_TOKEN':
+          return {
+            ...prevState,
+            userToken: action.token,
+            isLoading: false,
+          };
+        case 'SIGN_IN':
+          return {
+            ...prevState,
+            isSignout: false,
+            userToken: action.token,
+          };
+        case 'SIGN_OUT':
+          return {
+            ...prevState,
+            isSignout: true,
+            userToken: null,
+          };
+        default:
+          return prevState;
+      }
+    },
+    {
+      isLoading: true,
+      isSignout: false,
+      userToken: null,
+    },
+  );
+
+export const getAuthContext = (
+  dispatch: React.Dispatch<AuthContextAction>,
+): AuthContextType => ({
+  signIn: async (email: string, password: string) => {
+    const auth = getAuth();
+    signInWithEmailAndPassword(auth, email, password)
+      .then(async userCredential => {
+        const { user } = userCredential;
+
+        console.log(
+          'Auth Success: signed in user with email',
+          userCredential.user.email,
+        );
+        // After getting the user credential, persist the uid in AsyncStorage
+        await AsyncStorage.setItem('uid', user.uid);
+        dispatch({ type: 'SIGN_IN', token: user.uid });
+      })
+      .catch(error => {
+        console.warn('(signIn) error', error);
+      });
+  },
+  signOut: async () => {
+    // Remove the uid from AsyncStorage
+    await AsyncStorage.removeItem('uid');
+    dispatch({ type: 'SIGN_OUT' });
+  },
+  signUp: async (email: string, password: string) => {
+    const auth = getAuth();
+    createUserWithEmailAndPassword(auth, email, password)
+      .then(async userCredential => {
+        const { user } = userCredential;
+        console.log(
+          'Auth Success: created user with email',
+          userCredential.user.email,
+        );
+        await AsyncStorage.setItem('uid', user.uid);
+        dispatch({ type: 'SIGN_IN', token: user.uid });
+      })
+      .catch(error => {
+        console.warn('(signUp) error', error);
+      });
+  },
+});
